+++ src/Module_Nodes/Module_Nodes.cpp
@@ -69,17 +69,15 @@ void Module_Nodes::handle_error(Cool::OptionalErrorMessage const& maybe_err, boo
             return make_shader_compilation_error_message(name(), "", msg);
         });
     }
-#if DEBUG
     else
     {
         maybe_err.send_error_if_any(
             [&](const std::string& msg) {
                 return make_shader_compilation_error_message("Test Nodes", "", msg);
             },
-            Cool::Log::Debug::console()
+            Cool::Log::ToUser::console()
         );
     }
-#endif
 }
 
 auto Module_Nodes::nodes_config(Ui_Ref ui) const -> NodesConfig
@@ -92,14 +90,12 @@ void Module_Nodes::imgui_windows(Ui_Ref ui) const
     if (_nodes_editor.imgui_window(nodes_config(ui), _nodes_library))
         ui.set_dirty(_regenerate_code_flag);
 
-#if DEBUG
     if (DebugOptions::show_generated_shader_code())
     {
         ImGui::Begin("Nodes Code");
         ImGui::InputTextMultiline("##Nodes shader code", &_shader_code, ImVec2{ImGui::GetWindowWidth() - 10, ImGui::GetWindowSize().y - 35});
         ImGui::End();
     }
-#endif
 }
 
 auto Module_Nodes::all_inputs() const -> Cool::AllInputRefsToConst
@@ -168,10 +164,8 @@ void Module_Nodes::render(RenderParams in, UpdateContext_Ref update_ctx)
 
     if (in.is_dirty(_regenerate_code_flag))
     {
-#if DEBUG
         if (DebugOptions::log_when_compiling_nodes())
-            Cool::Log::Debug::info("Nodes", "Compiled");
-#endif
+            Cool::Log::ToUser::info("Nodes", "Compiled");
         compile(update_ctx);
         in.set_clean(_regenerate_code_flag);
     }