+++ changelog.md
@@ -12,7 +12,8 @@
   - Presets for all of them
   - Reorganized categories
 - Show debug options
-- More info in the Info menu (OS and GPU version)
+- Moved debug info from the Info menu to the Debug menu
+- More info in the Debug menu (OS and GPU version)
 - Added possibility to generate an info dump to help developers debug
 
 ## beta-0